DST-1136 Typeahead market filters with vanilla js
Vue version of the tyeahead seems to have problems with cross browser compatibility because
we're relying on both browser behaviour and vue state. keeping both in sync has proved to be fiddly, thus a simpler approach was needed.

Implement a simple vanilla js filter that toggles a class to the dom elements that don't match the search query.