FEX-107 Add dissolved company notice on results page
A new function prepareCompany was added to transformers.js.
This contains the logic required to prepare the dissolved company
notice for the nunjucks template. In future, any additional logic
required to transform the company result as returned by dt07
can be added to this function.

New config item daysUntilDissolvedCompaniesDeleted was added and set to
180.